+++ site/src/pages/calls.js
@@ -14,17 +14,15 @@ import {
   callListSelector,
   clearCallList,
 } from "../store/reducers/callSlice";
-import { useCallsFilter } from "../utils/hooks/callsFilter";
-import Filter from "../components/filter";
 import omit from "lodash.omit";
 import * as queryString from "query-string";
+import CallFilter from "../components/calls/filter";
 
 function Calls() {
   const location = useLocation();
   const dispatch = useDispatch();
   const page = getPageFromQuery(location);
   const pageSize = LIST_DEFAULT_PAGE_SIZE;
-  const filters = useCallsFilter();
 
   const list = useSelector(callListSelector);
   const loading = useSelector(callListLoadingSelector);
@@ -53,7 +51,7 @@ function Calls() {
   return (
     <Layout>
       <BreadCrumb data={[{ name: "Calls" }]} />
-      <Filter data={filters} />
+      <CallFilter />
 
       <StyledPanelTableWrapper
         footer={